Fixed warnings with losing precision

+++ trunk/Bricksmith/Source/Application/General/DonationDialogController.m
@@ -117,7 +117,7 @@
 	CFBundleRef     mainBundle                  = CFBundleGetMainBundle();
 	UInt32          bundleVersion               = CFBundleGetVersionNumber(mainBundle);
 	BOOL            userRequestedSuppression    = [userDefaults boolForKey:DONATION_SCREEN_SUPPRESS_THIS_VERSION];
-	UInt32          lastNagVersion              = [userDefaults integerForKey:DONATION_SCREEN_LAST_VERSION_DISPLAYED];
+	NSInteger       lastNagVersion              = [userDefaults integerForKey:DONATION_SCREEN_LAST_VERSION_DISPLAYED];
 	BOOL            showDonationRequest         = YES;
 	
 	if(userRequestedSuppression == YES)
